+++ app/(auth)/login.tsx
@@ -1,23 +1,24 @@
+import Button from '@/components/Atoms/Button';
+import PhoneInput from '@/components/Atoms/PhoneInput';
+import { ThemedText } from '@/components/ThemedText';
+import { Spacing } from '@/constants/Theme';
+import { useAuth } from '@/context/AuthContext';
+import { useTheme } from '@/context/ThemeContext';
+import { router } from 'expo-router';
+import { ChevronLeft } from 'lucide-react-native';
 import React, { useState } from 'react';
 import {
-  View,
-  Text,
-  StyleSheet,
   KeyboardAvoidingView,
   Platform,
-  TouchableOpacity,
   ScrollView,
+  StyleSheet,
+  TouchableOpacity,
+  View,
 } from 'react-native';
-import { router } from 'expo-router';
-import { useAuth } from '@/context/AuthContext';
-import PhoneInput from '@/components/Atoms/PhoneInput';
-import Button from '@/components/Atoms/Button';
-import Colors from '@/constants/Colors';
-import { ChevronLeft } from 'lucide-react-native';
-import { Spacing } from '@/constants/Theme';
 
 export default function LoginScreen() {
   const { login } = useAuth();
+  const { colors } = useTheme(); // Get dynamic theme colors
   const [phoneNumber, setPhoneNumber] = useState('');
   const [error, setError] = useState('');
   const [isLoading, setIsLoading] = useState(false);
@@ -35,12 +36,9 @@ export default function LoginScreen() {
 
     setIsLoading(true);
 
-    // Simulate API call
     setTimeout(() => {
       setIsLoading(false);
-      // Store the phone number in auth context
       login(phoneNumber);
-      // Navigate to OTP verification screen
       router.push({
         pathname: '/verify-otp',
         params: { mode: 'login', phone: phoneNumber },
@@ -50,7 +48,7 @@ export default function LoginScreen() {
 
   return (
     <KeyboardAvoidingView
-      style={styles.container}
+      style={[styles.container, { backgroundColor: colors.background }]}
       behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
       keyboardVerticalOffset={Platform.OS === 'ios' ? 64 : 0}
     >
@@ -60,18 +58,23 @@ export default function LoginScreen() {
       >
         <View style={styles.header}>
           <TouchableOpacity
-            style={styles.backButton}
+            style={[styles.backButton, { backgroundColor: colors.border }]}
             onPress={() => router.back()}
           >
-            <ChevronLeft size={24} color={Colors.light.text} />
+            <ChevronLeft size={24} color={colors.text} />
           </TouchableOpacity>
-          <Text style={styles.title}>Log in</Text>
+          <ThemedText
+            style={[styles.title, { color: colors.text }]}
+            type="title"
+          >
+            Log in
+          </ThemedText>
         </View>
 
         <View style={styles.formContainer}>
-          <Text style={styles.subtitle}>
+          <ThemedText style={[styles.subtitle, { color: colors.darkGray }]}>
             Enter your phone number to receive a verification code
-          </Text>
+          </ThemedText>
 
           <PhoneInput
             value={phoneNumber}
@@ -93,20 +96,24 @@ export default function LoginScreen() {
             style={styles.forgotPasswordLink}
             onPress={() => router.push('/forgot-password')}
           >
-            <Text style={styles.forgotPasswordText}>Forgot Password?</Text>
+            <ThemedText
+              style={[styles.forgotPasswordText, { color: colors.primary }]}
+            >
+              Forgot Password?
+            </ThemedText>
           </TouchableOpacity>
         </View>
 
         <View style={styles.footer}>
-          <Text style={styles.footerText}>
+          <ThemedText style={[styles.footerText, { color: colors.darkGray }]}>
             Don't have an account?{' '}
-            <Text
-              style={styles.signupLink}
+            <ThemedText
+              style={[styles.signupLink, { color: colors.primary }]}
               onPress={() => router.push('/signup')}
             >
               Sign up
-            </Text>
-          </Text>
+            </ThemedText>
+          </ThemedText>
         </View>
       </ScrollView>
     </KeyboardAvoidingView>
@@ -116,7 +123,6 @@ export default function LoginScreen() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: Colors.light.background,
   },
   scrollContainer: {
     flexGrow: 1,
@@ -130,20 +136,15 @@ const styles = StyleSheet.create({
     width: 40,
     height: 40,
     borderRadius: 20,
-    backgroundColor: 'rgba(0, 0, 0, 0.05)',
     justifyContent: 'center',
     alignItems: 'center',
     marginBottom: 16,
   },
   title: {
     fontFamily: 'PlusJakartaSans-Bold',
-    fontSize: 32,
-    color: Colors.light.text,
   },
   subtitle: {
     fontFamily: 'Inter-Regular',
-    fontSize: 16,
-    color: Colors.light.darkGray,
     marginBottom: 24,
     lineHeight: 22,
   },
@@ -162,8 +163,6 @@ const styles = StyleSheet.create({
   },
   forgotPasswordText: {
     fontFamily: 'Inter-Medium',
-    fontSize: 16,
-    color: Colors.light.primary,
   },
   footer: {
     marginTop: 'auto',
@@ -172,11 +171,8 @@ const styles = StyleSheet.create({
   },
   footerText: {
     fontFamily: 'Inter-Regular',
-    fontSize: 16,
-    color: Colors.light.darkGray,
   },
   signupLink: {
     fontFamily: 'Inter-SemiBold',
-    color: Colors.light.primary,
   },
 });
