+++ app/_layout.tsx
@@ -61,7 +61,6 @@ export default function RootLayout() {
           <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
           <Stack.Screen name="+not-found" options={{ title: 'Oops!' }} />
         </Stack>
-        <StatusBar style="auto" />
       </AuthProvider>
     </ThemeProvider>
   );
